refactor: PR #344 Round 14 리뷰 코멘트 반영 (도메인 타입 중앙화, useDialog 상태 통일, cardStyles 주석 정확화)
- domain/types.ts: SkillStat, SkillStatConfig 인터페이스 추가 (MISTAKES.md Architecture #1)
- domain/skills.ts: 로컬 인터페이스 정의 제거, domain/types에서 import
- useDialog.ts: 중복된 close 로직 통일 (setIsOpen + focus restore → 단일 useCallback)
- cardStyles.ts: IndexCard와 SignalStockCard 사용 패턴 반영한 주석 수정
