+++ src/domain/types.ts
@@ -799,3 +799,12 @@ export type QuadrantKey =
     | 'bullishExpected'
     | 'bearishExpected'
     | 'bearishConfirmed';
+
+export interface SkillStat {
+    value: number;
+    label: string;
+}
+
+export interface SkillStatConfig {
+    countLabel: string;
+}
